added in a check to see if ajenti is installed and if it is done add a repo or try to install.

+++ install/Debian/Debian-Wheezy-PBX-PKG-Install-new.sh
@@ -740,11 +740,13 @@ DELIM
 fi
 
 #ADD Ajenti repo
+if [[ -f /etc/ajenti ]]; then
 /bin/cat > "/etc/apt/sources.list.d/ajenti.list" <<DELIM
 deb http://repo.ajenti.org/debian main main debian
 DELIM
 wget http://repo.ajenti.org/debian/key -O- | apt-key add -
 apt-get update &> /dev/null
+fi
 
 #Install ajenti openvpn openvpn-scripts pbx-admin-menu
 for i in ajenti openvpn openvpn-scripts pbx-admin-menu ;do apt-get -y install --force-yes "${i}" ; done
